removed IO with Files (unnecesarry). And fixed warnings (mostly type conversions)

+++ Mortify/src/Mortify/Core/Application.cpp
@@ -63,7 +63,7 @@ namespace Mortify
 		{
 			MT_PROFILE_SCOPE("RunLoop");
 			
-			float time = m_OS->GetTime();
+			double time = m_OS->GetTime();
 			Timestep ts = time - m_TimeFromLastFrame;
 			m_TimeFromLastFrame = time;
